rename fdbutils -> fdbutil
