+++ setup.py
@@ -41,7 +41,7 @@ setup(
         'ssdutil',
         'pfc',
         'psuutil',
-        'fdbutils',
+        'fdbutil',
         'fwutil',
         'pcieutil',
         'pddf_fanutil',
@@ -123,7 +123,7 @@ setup(
             'counterpoll = counterpoll.main:cli',
             'crm = crm.main:cli',
             'debug = debug.main:cli',
-            'filter_fdb_entries = fdbutils.filter_fdb_entries:main',
+            'filter_fdb_entries = fdbutil.filter_fdb_entries:main',
             'pfcwd = pfcwd.main:cli',
             'sfputil = sfputil.main:cli',
             'ssdutil = ssdutil.main:ssdutil',
