+++ tests/filter_fdb_entries_test.py
@@ -7,7 +7,7 @@ import subprocess
 
 from collections import defaultdict
 from filter_fdb_input.test_vectors import filterFdbEntriesTestVector
-from fdbutils.filter_fdb_entries import main as filterFdbMain
+from fdbutil.filter_fdb_entries import main as filterFdbMain
 
 class TestFilterFdbEntries(object):
     """
